+++ sym/src/chacha20_ssse3.c
@@ -17,29 +17,30 @@
  * DAMAGE.
  */
 
+
 #include "chacha20_internal.h"
-#include "chacha20_neon.h"
+#include "chacha20_ssse3.h"
 #include "lc_chacha20.h"
 #include "lc_chacha20_private.h"
 #include "lc_sym.h"
 #include "visibility.h"
 
-#include "asm/ARMv8/chacha20_asm_neon.h"
+#include "asm/SSSE3/chacha20_arm_ssse3.h"
 
-static void cc20_crypt_neon(struct lc_sym_state *ctx, const uint8_t *in,
-			   uint8_t *out, size_t len)
+static void cc20_crypt_ssse3(struct lc_sym_state *ctx, const uint8_t *in,
+			     uint8_t *out, size_t len)
 {
-	cc20_crypt_asm(ctx, in, out, len, ChaCha20_neon);
+	cc20_crypt_asm(ctx, in, out, len, ChaCha20_ssse3);
 }
 
-static struct lc_sym _lc_chacha20_neon = {
+static struct lc_sym _lc_chacha20_ssse3 = {
 	.init = cc20_init,
 	.setkey = cc20_setkey,
 	.setiv = cc20_setiv,
-	.encrypt = cc20_crypt_neon,
-	.decrypt = cc20_crypt_neon,
+	.encrypt = cc20_crypt_ssse3,
+	.decrypt = cc20_crypt_ssse3,
 	.statesize = LC_CC20_BLOCK_SIZE,
 	.blocksize = 1,
 };
-LC_INTERFACE_SYMBOL(const struct lc_sym *, lc_chacha20_neon) =
-	&_lc_chacha20_neon;
+LC_INTERFACE_SYMBOL(const struct lc_sym *, lc_chacha20_ssse3) =
+	&_lc_chacha20_ssse3;