+++ sym/src/chacha20_ssse3_null.c
@@ -1,7 +1,7 @@
 /*
- * Copyright (C) 2016 - 2025, Stephan Mueller <smueller@chronox.de>
+ * Copyright (C) 2025, Stephan Mueller <smueller@chronox.de>
  *
- * License: see COPYING file in root directory
+ * License: see LICENSE file in root directory
  *
  * THIS SOFTWARE IS PROVIDED ``AS IS'' AND ANY EXPRESS OR IMPLIED
  * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
@@ -17,29 +17,8 @@
  * DAMAGE.
  */
 
-#include "chacha20_internal.h"
-#include "chacha20_neon.h"
+#include "chacha20_ssse3.h"
 #include "lc_chacha20.h"
-#include "lc_chacha20_private.h"
-#include "lc_sym.h"
 #include "visibility.h"
 
-#include "asm/ARMv8/chacha20_asm_neon.h"
-
-static void cc20_crypt_neon(struct lc_sym_state *ctx, const uint8_t *in,
-			   uint8_t *out, size_t len)
-{
-	cc20_crypt_asm(ctx, in, out, len, ChaCha20_neon);
-}
-
-static struct lc_sym _lc_chacha20_neon = {
-	.init = cc20_init,
-	.setkey = cc20_setkey,
-	.setiv = cc20_setiv,
-	.encrypt = cc20_crypt_neon,
-	.decrypt = cc20_crypt_neon,
-	.statesize = LC_CC20_BLOCK_SIZE,
-	.blocksize = 1,
-};
-LC_INTERFACE_SYMBOL(const struct lc_sym *, lc_chacha20_neon) =
-	&_lc_chacha20_neon;
+LC_INTERFACE_SYMBOL(const struct lc_sym *, lc_chacha20_ssse3) = NULL;